Add authentication to server actions

+++ actions/profile/delete-profile.ts
@@ -1,11 +1,16 @@
 "use server";
 
+import { auth } from "@clerk/nextjs";
 import prisma from "../prisma";
 
 export default async function DeleteProfile(clerk: string) {
+  const { userId } = await auth();
+
+  if (!userId) return;
+
   await prisma.profile.delete({
     where: {
-      clerk: clerk,
+      clerk: userId,
     },
   });
 }
