+++ actions/project/delete-project.ts
@@ -2,7 +2,12 @@
 
 import prisma from "../prisma";
 
+import { auth } from "@clerk/nextjs";
+
 export default async function DeleteProject(clerk: string, id: string) {
+  const { userId } = await auth();
+  if (!userId) return;
+
   await prisma.project.delete({
     where: {
       id: id,
